[DROOLS-6941] isolate eclipse compiler from drools-core + refactor
