+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/ClassDefinitionFactory.java
@@ -46,8 +46,8 @@ import org.drools.core.base.ClassFieldInspector;
 import org.drools.core.base.CoreComponentsBuilder;
 import org.drools.core.factmodel.AnnotationDefinition;
 import org.drools.core.factmodel.ClassDefinition;
-import org.drools.core.factmodel.EnumClassDefinition;
-import org.drools.core.factmodel.EnumLiteralDefinition;
+import org.drools.compiler.builder.impl.classbuilder.EnumClassDefinition;
+import org.drools.compiler.builder.impl.classbuilder.EnumLiteralDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.kie.internal.definition.GenericTypeDefinition;
 import org.drools.core.factmodel.traits.Thing;
