+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/classbuilder/BeanClassBuilder.java
@@ -14,23 +14,7 @@
  * limitations under the License.
  */
 
-package org.drools.core.factmodel;
+package org.drools.compiler.builder.impl.classbuilder;
 
-import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-
-import org.kie.api.internal.utils.KieService;
-
-public interface ClassBuilder extends KieService {
-
-    byte[] buildClass( ClassDefinition def, ClassLoader classLoader ) throws IOException,
-            SecurityException,
-            IllegalArgumentException,
-            ClassNotFoundException,
-            NoSuchMethodException,
-            IllegalAccessException,
-            InvocationTargetException,
-            InstantiationException,
-            NoSuchFieldException;
-
+public interface BeanClassBuilder extends ClassBuilder {
 }
