+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/classbuilder/ClassBuilder.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2011 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.compiler.builder.impl.classbuilder;
+
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+
+import org.drools.core.factmodel.ClassDefinition;
+import org.kie.api.internal.utils.KieService;
+
+public interface ClassBuilder extends KieService {
+
+    byte[] buildClass(ClassDefinition def, ClassLoader classLoader ) throws IOException,
+            SecurityException,
+            IllegalArgumentException,
+            ClassNotFoundException,
+            NoSuchMethodException,
+            IllegalAccessException,
+            InvocationTargetException,
+            InstantiationException,
+            NoSuchFieldException;
+
+}
