+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/classbuilder/EnumClassBuilder.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2010 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2011 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,25 +14,7 @@
  * limitations under the License.
  */
 
-package org.drools.core.rule;
+package org.drools.compiler.builder.impl.classbuilder;
 
-import org.drools.core.base.FieldAccessor;
-
-/**
- * An interface for Fact Fields
- */
-public interface FactField
-    extends
-    java.io.Externalizable {
-
-    public Class< ? > getType();
-
-    public String getName();
-
-    public boolean isKey();
-
-    public FieldAccessor getFieldAccessor();
-
-    public int getIndex();
-
+public interface EnumClassBuilder extends ClassBuilder {
 }
