+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/classbuilder/EnumClassDefinition.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.drools.core.factmodel;
+package org.drools.compiler.builder.impl.classbuilder;
 
 import java.io.IOException;
 import java.io.ObjectInput;
@@ -22,6 +22,8 @@ import java.io.ObjectOutput;
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Collections;
+
+import org.drools.core.factmodel.ClassDefinition;
 
 /**
  * Declares an enum class to be dynamically created
