+++ drools-ecj/src/main/java/org/drools/ecj/EclipseJavaCompiler.java
@@ -24,9 +24,9 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.StringTokenizer;
 
-import org.drools.core.factmodel.ClassBuilderFactory;
 import org.drools.util.ClassUtils;
 import org.drools.util.IoUtils;
+import org.drools.util.PortablePath;
 import org.eclipse.jdt.core.compiler.IProblem;
 import org.eclipse.jdt.internal.compiler.ClassFile;
 import org.eclipse.jdt.internal.compiler.Compiler;
@@ -44,8 +44,8 @@ import org.eclipse.jdt.internal.compiler.problem.DefaultProblemFactory;
 import org.kie.memorycompiler.AbstractJavaCompiler;
 import org.kie.memorycompiler.CompilationProblem;
 import org.kie.memorycompiler.CompilationResult;
+import org.kie.memorycompiler.JavaCompiler;
 import org.kie.memorycompiler.JavaCompilerSettings;
-import org.drools.util.PortablePath;
 import org.kie.memorycompiler.resources.ResourceReader;
 import org.kie.memorycompiler.resources.ResourceStore;
 
@@ -379,7 +379,7 @@ public final class EclipseJavaCompiler extends AbstractJavaCompiler {
         
         final Compiler compiler = new Compiler(nameEnvironment, policy, compilerOptions, compilerRequestor, problemFactory);
 
-        if ( ClassBuilderFactory.DUMP_GENERATED_CLASSES ) {
+        if ( JavaCompiler.DUMP_GENERATED_CLASSES ) {
             dumpUnits( compilationUnits, pReader );
         }
