+++ drools-traits/src/main/java/org/drools/traits/core/factmodel/AbstractProxyClassBuilderImpl.java
@@ -25,7 +25,7 @@ import java.util.BitSet;
 import java.util.Iterator;
 import java.util.Map;
 
-import org.drools.core.factmodel.BuildUtils;
+import org.drools.compiler.builder.impl.classbuilder.BuildUtils;
 import org.drools.core.factmodel.ClassDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.drools.core.factmodel.traits.CoreWrapper;
