+++ drools-traits/src/main/java/org/drools/traits/core/factmodel/AbstractTraitFactory.java
@@ -31,7 +31,7 @@ import java.util.Map;
 
 import org.drools.mvel.accessors.ClassFieldAccessor;
 import org.drools.mvel.accessors.ClassFieldAccessorStore;
-import org.drools.core.factmodel.BuildUtils;
+import org.drools.compiler.builder.impl.classbuilder.BuildUtils;
 import org.drools.core.factmodel.ClassDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.drools.core.factmodel.traits.CoreWrapper;
