+++ drools-traits/src/main/java/org/drools/traits/core/factmodel/Jenerator.java
@@ -26,7 +26,7 @@ import java.util.jar.JarOutputStream;
 
 import com.thoughtworks.xstream.XStream;
 import com.thoughtworks.xstream.io.xml.DomDriver;
-import org.drools.core.factmodel.ClassBuilder;
+import org.drools.compiler.builder.impl.classbuilder.ClassBuilder;
 import org.drools.core.factmodel.ClassDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.drools.util.IoUtils;
