+++ drools-traits/src/main/java/org/drools/traits/core/factmodel/TraitMapPropertyWrapperClassBuilderImpl.java
@@ -27,7 +27,7 @@ import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
-import org.drools.core.factmodel.BuildUtils;
+import org.drools.compiler.builder.impl.classbuilder.BuildUtils;
 import org.drools.core.factmodel.ClassDefinition;
 import org.drools.core.factmodel.FieldDefinition;
 import org.drools.mvel.asm.AsmUtil;
