+++ drools-traits/src/main/resources/META-INF/services/org.drools.compiler.builder.impl.classbuilder.ClassBuilderFactory
@@ -1,0 +1,1 @@
+org.drools.traits.core.factmodel.TraitClassBuilderFactory
